Fix commit.File() gorutine leak
Commit.File() was leaking a goroutine because it was looping over
an iterator without closing its channel.

Now commit.File() calls the new Tree.File() method that searches
the file in the repository by trasversing the dir tree instead
of using the tree.Files() iterator.

This not only prevent the goroutine leak, but also speeds up
file searching.