Change pktline API so you can add payloads
The old pktline API only had one method: New, that receives the
payloads and return the new PktLine, that was an io.Reader.  This means
you have to prepare the contents beforehand, in a [][]byte and then
call the ctor to build the pktlines.

Now, the construction of the pktlines and the method to add payloads are
separated:

  New() // creates an empty PktLines
  AddFlush()
  Add(pp ...[]byte)
  AddString(pp ...string)

and a PktLines has a public member R, which is the io.Reader of the
pktlines added.