RFC: provide insight into the status of gogit
A project I'm working on uses gogit to power a git remote helper
(i.e., it gets invoked by git when you do `git clone
<protocol>://...`, etc from the command line).  On big repos, gogit
and our backing store are slow enough that having printed progress
indicators makes a huge difference to usability, and better allows us
to debug where slownesses might be.

This is a quick-and-dirty PR that lets gogit communicate its status
back to some listener.  I haven't fixed up the tests yet, or written
any new tests, because I wanted to get feedback first on if this is
even a good approach or not.  Please take a look and let me know!

(Note that this is different from the sideband progress stuff, which
only lets you know the progress of the remote server.  This PR gives
the status of the local client.)