plumbing: the commit walker can skip externally-seen commits
When the revlist is computing the set of hashes needed to transfer, it
doesn't need to walk over commits it has already processed.  So, it
can instruct the commit walker not to walk those commits by passing in
its own `seen` map.

For a 36K object repo, this brought the time for `revlist.Objects`
down from 50s to 30s.