remote: option to skip compression while encoding packfiles
One use of go-git is to transfer git data from a non-standard git repo
(not stored in a file system, for example) to a "remote" backed by a
standard, local .git repo

In this scenario, delta compression is not needed to reduce transfer
time over the "network", because there is no network.  The underlying
storage layer has already taken care of the data tranfer, and sending
the objects to local .git storage doesn't require compression.  So
this PR gives the user the option to turn off compression when it
isn't needed.

Of course, this results in a larger, uncompressed local .git repo, but
the user can then run `git gc` or `git repack` on that repo if they
care about the storage costs.

Turning this on reduces total push time of a 36K repo by 50 seconds
(out of a pre-PR total of 3m26s).