Provide ability to add excludes to worktree
Adding excludes to the worktree seemed to be the best course of action since it minimizes
the amount of code changes; neatly sidestepping the issue of ballooning constructors.

It also seems to be philosophically in line w/ other code.  For example, there is no code to
obtain the username and email from the ~/.gitconfig file and yet those two pieces of
information are needed for commits, etc.

With that said, convenience functions, LoadGlobalPatterns() and LoadGlobalPatterns(),
are provided to obtain any excludes that may be declared in ~/.gitconfig and /etc/gitconfig
respectively.

Signed-off-by: Alan D. Cabrera <adc@toolazydogs.com>